backend/requirements.txt
# Framework web
Flask==2.3.3
Flask-CORS==4.0.0

# Autenticação e segurança
PyJWT==2.8.0

# Configuração
python-dotenv==1.0.0

# HTTP requests (para OAuth)
requests==2.31.0

# WSGI server para produção
gunicorn==21.2.0

# Utilitários
Werkzeug==2.3.7

# Dependências de teste
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0